% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthrop.R
\docType{data}
\name{anthrop}
\alias{anthrop}
\title{Length of Left Middle Finger and Height for 3000 Criminals}
\format{
Data frame with the following 2 variables: 
\describe{
  \item{finger}{length of left middle finger (cm)}
  \item{height}{height (inches)}
}
}
\source{
Macdonell, W. R. (1901). On criminal anthropometry and the
identification of criminals, \emph{Biometrika}, 1: 177--227.
}
\usage{
anthrop
}
\description{
Length of left middle finger and height for 3000 criminals
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
  438.
}
\keyword{datasets}
