% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\docType{data}
\name{audit}
\alias{audit}
\title{Selection of Accounts for Audit in Example 6.11}
\format{
Data frame with the following 6 variables: 
\describe{
  \item{account}{audit unit}
  \item{bookval}{book value of account}
  \item{cumbv}{cumulative book value}
  \item{rn1}{random number 1 selecting account}
  \item{rn2}{random number 2 selecting account}
  \item{rn3}{random number 3 selecting account}
}
}
\usage{
audit
}
\description{
Selection of Accounts for Audit in Example 6.11
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
  439.
}
\keyword{datasets}
