% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{books}
\alias{books}
\title{Data from Home Owner's Survey on Total Number of Books}
\format{
Data frame with the following 6 variables: 
\describe{
  \item{shelf}{shelf number}
  \item{number}{number of the book selected}
  \item{purchase}{purchase cost of the book}
  \item{replace}{replacement cost of book}
}
}
\usage{
books
}
\description{
Data from home owner's survey on total number of books
}
\note{
Used in Exercise 6 of Chapter 5.
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
  439.
}
\keyword{datasets}
