% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htpop.R
\docType{data}
\name{htpop}
\alias{htpop}
\title{Height and gender of 2000 persons in an artificial population}
\format{
\describe{
  \item{height}{height of person, cm}
  \item{gender}{factor with levels \code{F} (female) and \code{M} (male)}
}
}
\usage{
htpop
}
\description{
Height and gender of 2000 persons in an artificial population
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. 230--234 and
  441.
}
\keyword{datasets}
