% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsrs.R
\docType{data}
\name{htsrs}
\alias{htsrs}
\title{Height and gender for an SRS of 200 persons, taken from htpop}
\format{
\describe{
  \item{rn}{random number used to select the unit}
  \item{height}{height of person, cm}
  \item{gender}{factor with levels \code{F} (female) and \code{M} (male)}
}
}
\usage{
htsrs
}
\description{
Height and gender for an SRS of 200 persons, taken from htpop
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. 230--234 and
  442.
}
\keyword{datasets}
