% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htstrat.R
\docType{data}
\name{htstrat}
\alias{htstrat}
\title{Height and gender for a stratified random sample from htpop}
\format{
\describe{
  \item{rn}{random number used to select the unit}
  \item{height}{height of person, cm}
  \item{gender}{factor with levels \code{F} (female) and \code{M} (male)}
}
}
\usage{
htstrat
}
\description{
Height and gender for a stratified random sample of 160 women
and 40 men taken from the htpop population
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. 230--234 and
  442.
}
\keyword{datasets}
