% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\docType{data}
\name{journal}
\alias{journal}
\title{Types of Sampling Used for Articles in a Sample of Journals}
\format{
Data frame with the following 3 variables: 
\describe{
  \item{numemp}{number of articles in 1988 that used sampling}
  \item{prob}{number of articlues that used probability sampling}
  \item{nonprob}{number of articles that used nonprobability sampling}
}
}
\source{
Jacoby and Handlin (1991). TODO
}
\usage{
journal
}
\description{
Types of Sampling Used for Articles in a Sample of Journals
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
  442.
}
\keyword{datasets}
