% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lahiri.design.R
\name{lahiri.design}
\alias{lahiri.design}
\title{Draw Samples Using Lahiri's Method}
\usage{
lahiri.design(relsize, n, clnames = seq(along = relsize))
}
\arguments{
\item{relsize}{vector of relative sizes of population PSUs}

\item{n}{desired sample size}

\item{clnames}{vector of PSU names for population}
}
\value{
clusters vector of n PSUs selected with replacement
        and with probability proportional to relsize
}
\description{
Draw Samples Using Lahiri's Method
}
\note{
Original code from Lohr (1999), p. 452 -- 453.
}
\references{
Lahiri, D. B. (1951). A method of sample selection 
            providing unbiased ratio estimates, \emph{Bulletin of
            the International Statistical Institute}, 33: 133 -- 140.
}
\author{
Sharon Lohr, slightly modified by Tobias Verbeke
}
