% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otters.R
\docType{data}
\name{otters}
\alias{otters}
\title{Otters Data}
\format{
Data frame with the following three variables: 
\describe{
  \item{section}{coastline section}
  \item{habitat}{type of habitat (stratum)}
  \item{holts}{number of holts}
}
}
\source{
Kruuk, H.A. et al. (1989). An estimate of numbers and habitat
  preferences  of otters \emph{Lutra lutra} in Shetland, UK., 
  Biological Conservation, 49: 241--254.
}
\usage{
otters
}
\description{
Data on number of holts (dens) in Shetland, United Kingdom used
in Kruuk et al. (1989). (Data courtesy of Hans Kruuk).
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
  443.
}
\keyword{datasets}
