% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seals.R
\docType{data}
\name{seals}
\alias{seals}
\title{Breathing Holes of Seals}
\format{
Data frame with the following 2 variables: 
\describe{
  \item{zone}{zone number for sampled area}
  \item{holes}{number of breathing holes Imjak found in area}
}
}
\source{
Lydersen, C. and Ryg, M. (1991). Evaluating breeding habitat and
populations of ringed seals \emph{Phoca hispida} in Svalbard fjords, 
\emph{Polar Record}, 27: 223--228.
}
\usage{
seals
}
\description{
Data on number of breathing holes found in sampled areas of Svalbard 
fjords, reconstructed from summary statistics given in Lydersen and
Ryg (1991)
}
\note{
The data are used in Chapter 4, Exercise 11.
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
  444.
}
\keyword{datasets}
