% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepps.R
\docType{data}
\name{statepps}
\alias{statepps}
\title{Information on States}
\format{
Date frame with the following 7 variables:
\describe{
  \item{state}{state name}
  \item{counties}{number of counties in state}
  \item{cumcount}{cumulative number of counties}
  \item{landarea}{land area of state, 1990 (square miles)}
  \item{cumland}{cumulative land area}
  \item{popn}{population of state, 1992}
  \item{cumpopn}{cumulative population}
}
}
\usage{
statepps
}
\description{
Number of counties, land area, and population for the 50 states plus
the District of Columbia
}
\references{
Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
  445.
}
\keyword{datasets}
