% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachnr.R
\docType{data}
\name{teachnr}
\alias{teachnr}
\title{Follow-Up Study of Nonrespondents from Gnap (1995)}
\format{
data frame with the following 6 variables:
\describe{
  \item{hrwork}{number of hours required to work at school per week}
  \item{size}{class size}
  \item{preprmin}{minutes spent per week in school on preparation}
  \item{assist}{minutes per week that a teacher's aide works with
    the teacher in the classroom}
}
}
\source{
Data courtesy of Rita Gnap (1995).
}
\usage{
teachnr
}
\description{
Follow-up study of nonrespondents from the Gnap (1995) study on the
workload of elementary school teachers in Maricopa County, Arizona.
}
\note{
The study is described in Exercise 16 of Chapter 15. The 
actual date are given in \code{\link{teachers}}. Cluster size data for
the original study are given in \code{\link{teachmi}}.
}
\references{
Gnap, R. (1995). Teacher load in Arizona elementary school 
districts in Maricopa County. Ph.D. diss., Arizona State University 

Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
  446.
}
\keyword{datasets}
